S01E02
Hello and Welcome to Olympus Mons Tutorials!

In this episode we will add a few more things such as a Window Process and Icon files. At the end of this episode we will have a much better looking and efficient program.

This series will go through different aspects of c++ Win32, DirectX 12 Graphics API and many more as we strive to build a C++ DirectX Ready Game Engine.

Github Link:
https://github.com/OlympusMonsTutorials/OlympusMons

Please subscribe and like

And feel free to drop a comment!